Add files via upload
This is a basic HTML file with some CSS and JavaScript. The HTML sets up a header with a navigation menu on either side with links to the user's Github and LinkedIn profiles, as well as a main body that contains a button that when clicked will display an alert with the text "GET A LIFE ... LOSER!". The CSS styles the header, button and sets the background color of the body. The JavaScript uses event listeners to attach the "runAway" and "alertClick" functions to the button. The "runAway" function changes the position of the button randomly every time it's moused over and the "alertClick" function displays the alert message when the button is clicked.